Reslove bug and refactoring of tour implementation.
Changes include
0) Corrected the arrow coloring of the previous pointer and next pointer for the first step and the last step in the react tour.
1) Created a Tour Component under the Common/Tour folder.
2) Moved the ReactTour Component to Tour component under the Common/Tour folder.
3) Moved the ControlScroll.js from Common/Tour folder.
4) ran the "npm run lint" to fix the lint issue as suggested by @daltonfury42
due to this command most of the files are ending has changed from LF to CRLF.
